refactor: :unicorn_face: ディレクトリ名の変更
lib が不適切だったため、utils に変更
